+++ install-system.bash
@@ -14,7 +14,7 @@ set -e;  # Fail on each error
 
 # Install git if not installed yet
 if ! command -v git >/dev/null 2>&1; then
-    apt-get install -y git;
+    apt-get install -y git wget;
 fi
 
 # define vars